pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ym2s"
dynamic = ["version"]
description = 'Yandex Music to Spotify syncer'
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "Dmitry Starov", email = "dmitry.staroff@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
  "Typing :: Typed",
]
dependencies = [
  # CLI
  "cloup~=3.0.5",
  "click~=8.1.7",
  "inflect~=7.2.0",
  "colorama~=0.4.6",
  "humanfriendly~=10.0",
  # Internals
  "PyYAML~=6.0.1",
  "requests~=2.31.0",
  "mustopt~=1.0.1",
  # Legacy
  "Pillow",
  "spotipy~=2.23.0",
  "yandex-music~=2.2.0",
]

[project.urls]
Documentation = "https://github.com/dstaroff/ym2s#readme"
Issues = "https://github.com/dstaroff/ym2s/issues"
Source = "https://github.com/dstaroff/ym2s"

[project.scripts]
ym2s = "ym2s.cli:main"

[tool.mypy]
ignore_missing_imports = true
disallow_incomplete_defs = true
check_untyped_defs = true
exclude = ['tests/staging']
